# symbolic.py: faster Node.sum, faster SumNode.div