Minor refactor of symbolic.py

1) Simplify node partitioning in `Node.sum`
2) Replace `isinstance` with faster `__class__` check where possible.
3) Replace use of partition helper function with faster oneliner

I believe the timing gains of the first two are of the order of the measurement noise, the last change is however quite a bit faster.

Measured using `kernprof -lv pytest -s test/test_speed_v_torch.py`

`Node.sum`
Before
```
Total time: 0.715114 s
File: tinygrad/tinygrad/shape/symbolic.py
Function: sum at line 60

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
    60                                             @staticmethod
    61                                             @profile
    62                                             def sum(nodes:List[Node]) -> Node:
    63     25740       8056.8      0.3      1.1      new_nodes: List[Node] = []
    64     25740       6732.8      0.3      0.9      sum_nodes: List[SumNode] = []
    65     25740       6307.9      0.2      0.9      num_nodes: List[NumNode] = []
    66     25740       5825.5      0.2      0.8      mul_nodes: List[MulNode] = []
    67     91400      28279.7      0.3      4.0      for node in nodes:
    68     47923      17353.2      0.4      2.4        if isinstance(node, NumNode):
    69     43477      15765.7      0.4      2.2          num_nodes.append(node)
    70     37626      11982.1      0.3      1.7        elif isinstance(node, MulNode):
    71     37626      13254.2      0.4      1.9          mul_nodes.append(node)
    72      7837       2498.4      0.3      0.3        elif isinstance(node, SumNode): # expand any sums inside one sum
    73      2460        978.9      0.4      0.1          sum_nodes.append(node)
    74                                                 else:
    75      7837       2957.3      0.4      0.4          new_nodes.append(node)
    76                                           
    77                                               # expand any sums inside one sum
    78     24400       7164.3      0.3      1.0      if sum_nodes:
    79      1340        549.9      0.4      0.1        new_nodes.extend(num_nodes)
    80      1340        426.4      0.3      0.1        new_nodes.extend(mul_nodes)
    81      2460       1184.5      0.5      0.2        for x in sum_nodes: new_nodes += x.nodes
    82      1340       1688.4      1.3      0.2        return Variable.sum(new_nodes)
    83                                           
    84                                               # combine any numbers inside a sum
    85     22617       5451.7      0.2      0.8      if num_nodes:
    86     22617      41027.0      1.8      5.7        new_nodes.append(NumNode(sum([x.b for x in num_nodes])))
    87                                           
    88                                               # combine any MulNodes that factorize (big hack sticking the MulNode(x, 1) on things)
    89     24400      61250.7      2.5      8.6      mul_nodes += [MulNode(x, 1) for x in new_nodes]
    90     24400      14743.3      0.6      2.1      mul_groups: Dict[str, Tuple[Node, List[MulNode]]] = defaultdict(lambda: (Node(), []))
    91     67806      17835.3      0.3      2.5      for node in mul_nodes: #NOTE can we somehow avoid rendering here?
    92     67806      93136.9      1.4     13.0        key = node.a.render()
    93     67806      67779.6      1.0      9.5        mul_groups[key] = (node.a, mul_groups[key][1] + [node])
    94     24400     164845.8      6.8     23.1      mul_nodes = [k * sum(x.b for x in g) for k, g in mul_groups.values()]
    95     24400      32100.9      1.3      4.5      new_nodes = [x if not isinstance(x, MulNode) or x.b != 1 else x.a for x in mul_nodes]
    96                                           
    97                                               # filter 0s
    98     24400      29772.2      1.2      4.2      new_nodes = [x for x in new_nodes if x.min != 0 or x.max != 0]
    99     24400      56164.9      2.3      7.9      return create_rednode(SumNode, new_nodes) if len(new_nodes) > 1 else (new_nodes[0] if len(new_nodes) == 1 else NumNode(0))
```

After:
```
Total time: 0.713905 s
File: tinygrad/tinygrad/shape/symbolic.py
Function: sum at line 60

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
    60                                             @staticmethod
    61                                             @profile
    62                                             def sum(nodes:List[Node]) -> Node:
    63     25740      14202.0      0.6      2.0      new_nodes, num_nodes, mul_nodes, sum_nodes = [],[],[],[]
    64     25740      13887.1      0.5      1.9      lists = {NumNode : num_nodes, MulNode : mul_nodes, SumNode : sum_nodes}
    65     90790      25728.6      0.3      3.6      for node in nodes:
    66     90790      48282.1      0.5      6.8        lists.get(node.__class__, new_nodes).append(node)
    67                                           
    68                                               # expand any sums inside one sum
    69     24400       7378.9      0.3      1.0      if sum_nodes:
    70      1340        541.1      0.4      0.1        new_nodes.extend(num_nodes)
    71      1340        437.6      0.3      0.1        new_nodes.extend(mul_nodes)
    72      2460       1216.6      0.5      0.2        for x in sum_nodes: new_nodes += x.nodes
    73      1340       1965.5      1.5      0.3        return Variable.sum(new_nodes)
    74                                           
    75                                               # combine any numbers inside a sum
    76     22617       6167.3      0.3      0.9      if num_nodes:
    77     22617      42988.4      1.9      6.0        new_nodes.append(NumNode(sum([x.b for x in num_nodes])))
    78                                           
    79                                               # combine any MulNodes that factorize (big hack sticking the MulNode(x, 1) on things)
    80     24400      62585.3      2.6      8.8      mul_nodes += [MulNode(x, 1) for x in new_nodes]
    81     24400      15714.6      0.6      2.2      mul_groups: Dict[str, Tuple[Node, List[MulNode]]] = defaultdict(lambda: (Node(), []))
    82     67806      19338.9      0.3      2.7      for node in mul_nodes: #NOTE can we somehow avoid rendering here?
    83     67806      97178.8      1.4     13.6        key = node.a.render()
    84     67806      71796.2      1.1     10.1        mul_groups[key] = (node.a, mul_groups[key][1] + [node])
    85     24400     165823.0      6.8     23.2      mul_nodes = [k * sum(x.b for x in g) for k, g in mul_groups.values()]
    86     24400      32989.6      1.4      4.6      new_nodes = [x if not x.__class__ is MulNode or x.b != 1 else x.a for x in mul_nodes]
    87                                           
    88                                               # filter 0s
    89     24400      29561.6      1.2      4.1      new_nodes = [x for x in new_nodes if x.min != 0 or x.max != 0]
    90     24400      56121.9      2.3      7.9      return create_rednode(SumNode, new_nodes) if len(new_nodes) > 1 else (new_nodes[0] if len(new_nodes) == 1 else NumNode(0))
```


`__div__`
Before:
```
Total time: 0.76355 s
File: tinygrad/tinygrad/shape/symbolic.py
Function: __floordiv__ at line 169

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   169                                             @profile
   170                                             def __floordiv__(self, b: int, factoring_allowed=True):
   171     11932       3460.8      0.3      0.5      if not factoring_allowed: return Node.__floordiv__(self, b, factoring_allowed)
   172     11932      70917.6      5.9      9.3      factors, tmp_nofactor = partition(self.nodes, lambda x: (isinstance(x, (MulNode, NumNode))) and x.b%b == 0)
   173     11932       3137.5      0.3      0.4      nofactor = []
   174                                               # ugh, i doubt this is universally right
   175     30508       8548.6      0.3      1.1      for x in tmp_nofactor:
   176     19977       7129.3      0.4      0.9        if isinstance(x, NumNode):
   177      9921       3726.5      0.4      0.5          if (x.b%b) != x.b:
   178       610        899.4      1.5      0.1            factors.append(Variable.num(x.b - (x.b%b)))  # python does floor division
   179     10531      13414.6      1.3      1.8          nofactor.append(Variable.num(x.b%b))
   180                                                 else:
   181     19977       7202.3      0.4      0.9          nofactor.append(x)
   182     11932      22465.9      1.9      2.9      gcd = [math.gcd(x.b, b) if isinstance(x, (MulNode, NumNode)) else None for x in nofactor]
   183      7035       2796.9      0.4      0.4      if len(factors) > 0:
   184                                                 # these don't have to be the same, just having a common factor
   185      5758      13281.4      2.3      1.7        if len(gcd) > 0 and all_same(gcd) and gcd[0] is not None and gcd[0] > 1:
   186      1185      37760.1     31.9      4.9          nofactor_term = Variable.sum([(x.a * (x.b//gcd[0])) if isinstance(x, MulNode) else Variable.num(x.b//gcd[0]) for x in nofactor])//(b//gcd[0])
   187                                                 else:
   188      4029     203690.7     50.6     26.7          nofactor_term = Variable.sum(nofactor)//b
   189      7035     348552.9     49.5     45.6        return Variable.sum([(x.a * (x.b//b)) if isinstance(x, MulNode) else Variable.num(x.b//b) for x in factors] + [nofactor_term])
   190                                               else:
   191      4897       6141.9      1.3      0.8        muls = [x.b for x in nofactor if isinstance(x, MulNode)]
   192      4709       1376.6      0.3      0.2        for m in muls:
   193      4309       1747.2      0.4      0.2          if m > 1 and b%m == 0:
   194      4309       5289.6      1.2      0.7            return (self//m)//(b//m)
   195       588       2010.3      3.4      0.3        return Node.__floordiv__(self, b, factoring_allowed)
```

After:
```
Total time: 0.650139 s
File: tinygrad/tinygrad/shape/symbolic.py
Function: __floordiv__ at line 160

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   160                                             @profile
   161                                             def __floordiv__(self, b: int, factoring_allowed=True):
   162     11932       3383.3      0.3      0.5      if not factoring_allowed: return Node.__floordiv__(self, b, factoring_allowed)
   163     11932       3832.1      0.3      0.6      factors, tmp_nofactor, nofactor = [], [], []
   164     44852      31764.2      0.7      4.9      for x in self.nodes: factors.append(x) if x.__class__ in (MulNode, NumNode) and x.b%b == 0 else nofactor.append(x)**
   165                                               # ugh, i doubt this is universally right
   166     11932       3673.2      0.3      0.6      for x in tmp_nofactor:
   167                                                 if x.__class__ is NumNode:
   168                                                   if (x.b%b) != x.b:
   169                                                     factors.append(Variable.num(x.b - (x.b%b)))  # python does floor division
   170                                                   nofactor.append(Variable.num(x.b%b))
   171                                                 else:
   172                                                   nofactor.append(x)
   173     11932      22023.4      1.8      3.4      gcd = [math.gcd(x.b, b) if  x.__class__ in (MulNode, NumNode) else None for x in nofactor]
   174      7035       2873.0      0.4      0.4      if len(factors) > 0:
   175                                                 # these don't have to be the same, just having a common factor
   176      5758      14024.7      2.4      2.2        if len(gcd) > 0 and all_same(gcd) and gcd[0] is not None and gcd[0] > 1:
   177      1185      36181.0     30.5      5.6          nofactor_term = Variable.sum([(x.a * (x.b//gcd[0])) if x.__class__ is MulNode else Variable.num(x.b//gcd[0]) for x in nofactor])//(b//gcd[0])
   178                                                 else:
   179      4029     191465.6     47.5     29.4          nofactor_term = Variable.sum(nofactor)//b
   180      7035     323973.8     46.1     49.8        return Variable.sum([(x.a * (x.b//b)) if x.__class__ is MulNode else Variable.num(x.b//b) for x in factors] + [nofactor_term])
   181                                               else:
   182      4897       6274.3      1.3      1.0        muls = [x.b for x in nofactor if x.__class__ is MulNode]
   183      4709       1470.5      0.3      0.2        for m in muls:
   184      4309       1886.4      0.4      0.3          if m > 1 and b%m == 0:
   185      4309       5048.2      1.2      0.8            return (self//m)//(b//m)
   186       588       2265.7      3.9      0.3        return Node.__floordiv__(self, b, factoring_allowed)
```