@tinyb0t 2023-06-20
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/shape/symbolic.py                14              23
------------------------------------------------------------
lines added in the tinygrad folder: -9
```