@roelofvandijk 2023-06-21
New faster sum implementation. Change to draft MR.

- Loop over nodes once
- Add flatten property to SumNodes
- Only factoring when necessary
- Simplify factoring code

These functions get hit a lot and could be cached, but I think that the current hashing of nodes is slow and could be improved. Therefore most caching here is not actually speeding things up.

```
Total time: 0.482461 s
File: /home/rvd/src/roelofvandijk/tinygrad/tinygrad/shape/symbolic.py
Function: sum at line 70

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
    70                                             @staticmethod
    71                                             @profile
    72                                             def sum(nodes:List[Node]) -> Node:
    73     23565       7477.1      0.3      1.5      if not nodes: return NumNode(0)
    74     21407      10390.8      0.5      2.2      if len(nodes) == 1: return nodes[0]
    75                                           
    76     21407       6138.1      0.3      1.3      new_nodes: List[Node] = []
    77     21407       4953.4      0.2      1.0      a_nodes: List[Node] = []
    78     21407       5422.6      0.3      1.1      num_node_sum = 0
    79                                           
    80     85482      25509.5      0.3      5.3      for node in nodes:
    81     83022      32155.7      0.4      6.7        if isinstance(node, SumNode):
    82      2460      10407.0      4.2      2.2          flat = node.flat_components_grouped_num  # last component is always numnode
    83      2460       1327.0      0.5      0.3          nodes += flat[:-1]
    84      2460        924.1      0.4      0.2          num_node_sum += flat[-1].b
    85     45122      14828.4      0.3      3.1        elif node.__class__ is NumNode: 
    86     37900      12717.9      0.3      2.6          num_node_sum += node.b
    87                                                 else: 
    88     45122      17322.7      0.4      3.6          new_nodes.append(node)
    89     45122      22108.1      0.5      4.6          if isinstance(node, MulNode): a_nodes.append(node.a)
    90                                           
    91     11260       3064.0      0.3      0.6      if not num_node_sum:
    92      9317       2342.5      0.3      0.5        if not new_nodes: return NumNode(0)
    93      6465       3155.0      0.5      0.7        if len(new_nodes) == 1: return new_nodes[0]
    94     13516       3428.7      0.3      0.7      if not new_nodes: return NumNode(num_node_sum)
    95                                               
    96     13516     231653.4     17.1     48.0      if len(set(a_nodes)) < len(nodes): new_nodes = Node.factor(tuple(new_nodes)) # only factor if duplicate a-nodes exist
    97                                           
    98     13516      11530.4      0.9      2.4      if num_node_sum: new_nodes.append(NumNode(num_node_sum))  
    99                                           
   100     13516      55605.0      4.1     11.5      return create_rednode(SumNode, new_nodes) if len(new_nodes) > 1 else new_nodes[0] if len(new_nodes) == 1 else NumNode(0)
```