@roelofvandijk 2023-06-23
```
Total time: 0.222576 s
File: /home/rvd/src/roelofvandijk/tinygrad/tinygrad/shape/symbolic.py
Function: __floordiv__ at line 166

Line #      Hits         Time  Per Hit   % Time  Line Contents
==============================================================
   166                                             @profile
   167                                             def __floordiv__(self, b: int, factoring_allowed=True):
   168     11600       3523.2      0.3      1.6      if b == 1: return self
   169     11600       2902.6      0.3      1.3      if not factoring_allowed: return Node.__floordiv__(self, b, factoring_allowed)
   170     11600       2981.1      0.3      1.3      factors: List[Node] = []
   171     11600       2622.6      0.2      1.2      nofactor_mul: List[Node] = []
   172     11600       2690.3      0.2      1.2      nofactor_nonmul: List[Node] = []
   173     42755      30226.6      0.7     13.6      for x in self.flat_components: 
   174     42381      13452.0      0.3      6.0        if x.__class__ is NumNode and x.b%b == 0: factors.append(x)
   175     31130      17135.6      0.6      7.7        elif x.__class__ is MulNode: factors.append(x) if x.b%b == 0 else  nofactor_mul.append(x) 
   176     11251       3952.1      0.4      1.8        else: nofactor_nonmul.append(x)
   177                                                 
   178      6675       1694.7      0.3      0.8      if factors:  # factor out largest possible gcd
   179      6675      51426.2      7.7     23.1        factor_term = [x.a * x.b//b if isinstance(x, MulNode) else NumNode(x.b//b) for x in factors]
   180      6600       1830.5      0.3      0.8        if nofactor_mul and not nofactor_nonmul:
   181        75        115.3      1.5      0.1          gcds = [gcd(x.b, b) for x in nofactor_mul]
   182        75        158.3      2.1      0.1          if (t := min(gcds)) > 1 and all([x.b%t == 0 for x in nofactor_mul]): 
   183        51       1130.4     22.2      0.5            nofactor_term = [Node.sum([x.a * x.b//t for x in nofactor_mul if isinstance(x, MulNode)])//(b//t)]  # mypy wants the isinstance
   184                                                   else:
   185                                                     nofactor_term = [Node.sum(nofactor_mul)//b] if nofactor_mul else []
   186                                                 else: 
   187      4070      39456.4      9.7     17.7          nofactor_term = [Node.sum(nofactor_mul+nofactor_nonmul)//b] if nofactor_mul + nofactor_nonmul else []
   188      6675      38250.7      5.7     17.2        return Node.sum(factor_term + nofactor_term)
   189      4737       1270.8      0.3      0.6      for m in nofactor_mul:
   190      4337       5855.6      1.4      2.6        if m.b > 1 and b%m.b == 0: return (self//m.b)//(b//m.b)
   191       588       1900.9      3.2      0.9      return Node.__floordiv__(self, b, factoring_allowed)
```