@geohot 2023-06-23
Looks good! Minor nits in comments.

Btw, there's a fuzzer test for this: `python3 test/external/fuzz_symbolic.py` ran it and it looked good.