@roelofvandijk 2023-06-25
Thanks! Should be good to go. There's an efficient specialization for the case n=2, but I can add that later (if 2 different types and no SumNode -> create SumNode).

New timing should be around 0.2 for the function. Caching the factorization did not help so I left it out. I think that the current hashing of the nodes is not efficient, which makes caching operations here not efficient.