@roelofvandijk 2023-06-25
It would be very nice if there was a dedicated CI/CD test that evaluates a representative performance, e.g. for LLAMA.

For now I have been using test_net_speed, test_speed_v_torch and the llama example.