@geohot 2023-06-25
I would merge `test_speed_llama.py` that uses fake weights.