@geohot 2023-06-25
Profiler adds a lot of overhead, but with NOPROFILE=1 it's pretty close to real perf. #1046